% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.surv.R
\name{fplot.surv}
\alias{fplot.surv}
\title{Plot Kaplan-Meier Curves with number at risk and number censored at bottom}
\usage{
fplot.surv(smdl.yt = "Mths.OS", smdl.ye = "OS", smdl.x = "",
  xnames = "", xlab = "Months", ylab = "Overall Survival (\%)",
  dat = dat.work, t.at = NULL, lty = 1, mar = c(4, 3.5, 0.5, 0.5),
  mgp = c(2, 0.8, 0), nr.line = 3, nr.xadj = 0.75, nr.yadj = 0.8,
  loc.legend = "bottomleft", loc.p = "topright")
}
\arguments{
\item{smdl.yt}{survival time}

\item{smdl.ye}{event}

\item{smdl.x}{a covariate}

\item{xnames}{group names of the covariate x}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{dat}{data}

\item{t.at}{location of the x-axis tick marks}

\item{mar}{margin parameters}

\item{mgp}{mgp parameters}

\item{nr.line}{lines from the x-axis for the number at risk texts}

\item{nr.xadj}{adjustment along x-axis for the NR(NC) notation as a fraction of the length of the second tick mark from 0}

\item{nr.yadj}{adjustment along y-axis for the NR(NC) notation}

\item{loc.legend}{location of the legend}

\item{loc.p}{location of the p-value}
}
\value{
survfit output, median survival and coxph output if there is a covariate x
}
\description{
Plot Kaplan-Meier Curves with number at risk and number censored at bottom
}
\examples{
set.seed(16)
dat.work <- data.frame(StudyID=1:40,
                       Death = sample(c(0,1), 40, replace=T, p=c(0.2, 0.8)),
                       Mths.OS = 10*rexp(40, rate=0.5),
                       Group = factor(rep(c("grp1", "grp2"), c(15,25))))
out <- fplot.surv(smdl.yt = "Mths.OS",
                  smdl.ye = "Death",
                  xlab = "Months",
                  ylab = "Overall Survival (\%)",
                  dat = dat.work,
                  mar=c(4.0,3.5,0.5,0.7),
                  mgp = c(2.0, 0.8, 0),
                  nr.line = 3,
                  nr.xadj = 0.4,
                  nr.yadj = 0.8)

out <- fplot.surv(smdl.yt = "Mths.OS",
                  smdl.ye = "Death",
                  smdl.x = "Group",
                  xnames = c("Group1", "Group2"),
                  xlab = "Months",
                  ylab = "Overall Survival (\%)",
                  dat=dat.work,
                  t.at = NULL,
                  lty=NULL,
                  mar = c(4.0,3.5,0.5,0.7),
                  mgp = c(2.0, 0.8, 0),
                  nr.line = 3,
                  nr.xadj = 0.4,
                  nr.yadj = 0.8,
                  loc.legend = "bottomleft",
                  loc.p = "topright")

}
