% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmry.by.grp.R
\name{fsmry.by.grp}
\alias{fsmry.by.grp}
\title{Summarize a continuous variable by a categorical variable}
\usage{
fsmry.by.grp(y, grp, log.tr = F, method = "Tukey", method.padj = NULL,
  paired = F, IQR = T)
}
\arguments{
\item{y}{A continuous variable.}

\item{grp}{A categorical group variable.}

\item{log.tr}{An indicator of whether to log transform y}

\item{method}{Text string for multiple comparisons method if there are more than 2 categories, such as "Tukey", "Dunnet", or specifying the contrast}

\item{method.padj}{Text string for specifying method for multiple adjustment of p-values, e.g., "bonferroni", "fdr", etc.}

\item{paired}{An indicator of whether paired test should be used}

\item{IQR}{An indicator of whether to summarize y using IQR (default) or range.}
}
\value{
Summary of difference in y by grp using parametric and non-parametric methods.
}
\description{
Summarize a continuous variable by a categorical variable
}
\examples{
fsmry.by.grp(y=c(rnorm(10), rnorm(10, mean=0.5)),
             grp=factor(rep(c("grp1", "grp2"), each=10)))
}
