% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmry.graph.R
\name{fsmry.graph}
\alias{fsmry.graph}
\title{Summarize y by x graphically}
\usage{
fsmry.graph(y, x, type = c("scatter", "bxp", "errbar", "bar", "bar2"),
  y.plab = "", x.plab = "", ynames = NULL, xnames = NULL,
  stat.txt = NULL, loc.stat = NULL, ylim = NULL, fname = NULL,
  y.fnlab = NULL, x.fnlab = NULL, subgrp = "", withline = T,
  geomean = F, silent = T, width = 3, height = 3, mar = NULL,
  mgp = NULL, cex.bar = NULL, cex = 1, ...)
}
\arguments{
\item{y}{A continuous variable.}

\item{x}{A continuous or categorical variable.}

\item{type}{a text string for the type of graph, e.g., "scatter","bxp","errbar","bar",or "bar2"}

\item{y.plab}{y-axis label}

\item{x.plab}{x-axis label}

\item{ynames}{names for different categories of y}

\item{xnames}{names for different categories of x}

\item{stat.txt}{statistics to show on the graph, typically, p-value and/or correlation coefficient}

\item{loc.stat}{location of the stat.txt}

\item{ylim}{specify the y-axis limit if needed}

\item{fname}{a text string of file name if a win-meta file is to be generated}

\item{y.fnlab}{a text string of y variable name to be included in a file name}

\item{x.fnlab}{a text string of x variable name to be included in a file name}

\item{subgrp}{a text string of analysis cohort name to be included in a file name}

\item{withline}{a logic indicator of whether to include the Tukey line in the scatter plot, default true}

\item{geomean}{a logic indicator of whether to plot geometric mean/sd when plotting the errbar plot, default false}

\item{silent}{a logic indicator of whether to output the mp variable generated by barplot function, default true}

\item{width}{width of the graph is a win-meta file is to be generated}

\item{height}{height of the graph is a win-meta file is to be generated}

\item{mar}{graphical parameter}

\item{mgp}{graphical parameter}

\item{cex.bar}{graphical parameter for specifying text size of text above each bar}

\item{cex}{graphical parameter}

\item{...}{additional graphical paramter}
}
\value{
Summary graph of y by x.
}
\description{
Summarize y by x graphically
}
\examples{
set.seed(16)
dat.work <- data.frame(ht = c(rnorm(10, mean=1.72, sd=0.1), rnorm(10, mean=1.65, sd=0.1)),
                       wt = c(rnorm(10, mean=70, sd=10), rnorm(10, mean=60, sd=10)),
                       sex = factor(rep(c("Female", "Male", "Female", "Male"), c(2,8,6,4))),
                       group = factor(rep(c("grp1", "grp2"), each=10)))
cor.test(dat.work$wt, x = dat.work$ht)
fsmry.graph(y = dat.work$wt, x = dat.work$ht, type ="scatter",
            y.plab = "Weight",
            x.plab = "Height",
            stat.txt = "r=-0.24, p=0.31")
fsmry.by.grp(y = dat.work$ht, grp = dat.work$sex)
fsmry.graph(y = dat.work$ht, x = dat.work$sex, type ="bxp",
            y.plab = "Weight",
            x.plab = "Sex",
            stat.txt = "p=0.68")
fsmry.graph(y = dat.work$ht, x = dat.work$sex, type ="errbar",
            y.plab = "Weight",
            x.plab = "Sex",
            stat.txt = "p=0.66")
fsmry.graph(y = dat.work$ht, x = dat.work$sex, type ="errbar",
            y.plab = "Weight",
            stat.txt = "p=0.68",
            geomean=T)
fsmry2.by.grp(y = dat.work$sex, grp=dat.work$group, cmp.method="fisher")
fsmry.graph(y = dat.work$sex, x = dat.work$group, type ="bar",
            stat.txt = "p=0.17")
fsmry.graph(y = dat.work$sex, x = dat.work$group, type ="bar",
            xnames=c("Group1", "Group2"),
            stat.txt = "p=0.17")
fsmry.graph(y = dat.work$sex, x = dat.work$group, type ="bar2",
            xnames=c("Group1", "Group2"),
            stat.txt = "p=0.17")
}
