% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmry2.by.grp.R
\name{fsmry2.by.grp}
\alias{fsmry2.by.grp}
\title{Summarize a categorical variable by a categorical variable}
\usage{
fsmry2.by.grp(y, grp, cmp.method = c("fisher", "chisq"))
}
\arguments{
\item{y}{A categorical variable.}

\item{grp}{A categorical group variable.}

\item{cmp.method}{Text string for comparisons method, "fisher" or, "chisq"}

\item{An}{indicator of whether to log transform y}
}
\value{
Summary of difference in distribution of y by grp using Fisher's exact test or chi-squared test.
}
\description{
Summarize a categorical variable by a categorical variable
}
\examples{
fsmry2.by.grp(y = factor(rep(c("Female", "Male", "Female", "Male"), c(2,8,6,4))),
              grp = factor(rep(c("grp1", "grp2"), each=10)),
              cmp.method = "fisher")
}
